minor issues fixed

+++ lib/app/pages/whishlist/whishlist_controller.dart
@@ -1,6 +1,7 @@
 import 'package:coupon_app/app/base_controller.dart';
 import 'package:coupon_app/app/pages/whishlist/whishlist_presenter.dart';
 import 'package:coupon_app/app/utils/constants.dart';
+import 'package:coupon_app/app/utils/locale_keys.dart';
 import 'package:coupon_app/domain/entities/models/WhishlistItem.dart';
 
 class WhishlistController extends BaseController {
@@ -21,8 +22,11 @@ class WhishlistController extends BaseController {
   }
 
   delete(WhishlistItem item) {
-    showProgressDialog();
-    _presenter.deleteWhishlistItem(item);
+    showGenericConfirmDialog(getContext(), LocaleKeys.alert.tr(), LocaleKeys.confirmRemoveWhishlistItem.tr(), onConfirm: (){
+      whishListItems.remove(item);
+      refreshUI();
+      _presenter.deleteWhishlistItem(item);
+    });
   }
 
   void _initDeleteWhishlistItemListeners() {
